update: includes uploading images to storage from load-communities-to-firebase script
In order to achieve this, a google storage bucket has been created and set it as public
so any document can be accessed anonymously from a public URL. This can be done through
the google cloud console by adding 'allUsers' with reader role to the IAM control.

In the next steps community service would privide images with update url instead of
getting it from local files.

Related issue: #33